pyproject.toml
[project]
name = "vpn-daemon"
version = "0.4.0"
description = "Windows tray OpenVPN launcher with TOTP auth"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyotp>=2.9.0",
    "pystray>=0.19.5",
    "Pillow>=10.0.0",
    "keyring>=25.0",
]

[project.optional-dependencies]
helper = ["opencv-python-headless>=4.8"]

[project.scripts]
vpn-daemon = "vpn_daemon.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vpn_daemon"]

[dependency-groups]
dev = ["pytest>=8.0", "pytest-mock>=3.14"]
build = ["pyinstaller>=6.0"]
